Extract FP units to submodule
