+++ hw/chisel_acc/build.sbt
@@ -6,6 +6,8 @@ ThisBuild / organization := "be.kuleuven.esat.micas"
 
 val chiselVersion = "6.4.0"
 
+lazy val fpUnits = ProjectRef(file("subprojects/chisel-float"), "chiselFloat")
+
 lazy val root = (project in file("."))
   .settings(
     name := "snax-accelerators",
@@ -25,3 +27,4 @@ lazy val root = (project in file("."))
       "org.chipsalliance" % "chisel-plugin" % chiselVersion cross CrossVersion.full
     )
   )
+  .dependsOn(fpUnits, fpUnits % "compile->test")
