+++ hw/chisel_acc/project/build.properties
@@ -1,0 +1,1 @@
+sbt.version=1.11.3
