+++ hw/chisel_acc/project/plugins.sbt
@@ -1,3 +1,4 @@
 logLevel                    := Level.Warn
 addSbtPlugin("org.scalameta" % "sbt-scalafmt" % "2.5.2")
 addSbtPlugin("ch.epfl.scala" % "sbt-scalafix" % "0.11.1")
+addSbtPlugin("ch.epfl.scala" % "sbt-bloop"    % "1.5.11")
