+++ hw/chisel_acc/src/main/scala/snax_acc/versacore/VersaCore.scala
@@ -10,6 +10,8 @@ package snax_acc.versacore
 import chisel3._
 import chisel3.util._
 
+import fp_unit.FP16
+import fp_unit.FP32
 import snax_acc.utils._
 
 /** VersaCoreCfg is a configuration bundle for the VersaCore module. */
