+++ hw/chisel_acc/src/main/scala/snax_acc/versacore/VersaCoreGen.scala
@@ -13,6 +13,11 @@ import java.nio.file.Paths
 
 import chisel3._
 
+import fp_unit.DataType
+import fp_unit.FP16
+import fp_unit.FP32
+import fp_unit.FP8
+
 // hjson configuration parser, from hjson to SpatialArrayParam
 object SpatialArrayParamParser {
   def parseFromHjsonString(hjsonStr: String): SpatialArrayParam = {
